sorted ingredients now stay sorted after locking or unlocking

+++ app/Http/Controllers/IndexController.php
@@ -48,15 +48,40 @@ class IndexController extends Controller
         }
 
         // Energy, Proteins, Carbohydrates, Fat
+        $portionFactor = 2;
         $energy = [];
         $protein = [];
         $carbohydrate = [];
         $fat = [];
         foreach($ingredients as $ingredient){
-            array_push($energy,$ingredient->energy);
-            array_push($protein,$ingredient->protein);
-            array_push($carbohydrate,$ingredient->carbohydrate);
-            array_push($fat,$ingredient->fat);
+            // Topping 5% proportion
+            if($ingredient->category_id == 7){
+                array_push($energy,$ingredient->energy*0.05*$portionFactor);
+                array_push($protein,$ingredient->protein*0.05*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.05*$portionFactor);
+                array_push($fat,$ingredient->fat*0.05*$portionFactor);
+            }
+            // Salatbasis & Früchte 10% proportion
+            if($ingredient->category_id == 1 || $ingredient->category_id == 6){
+                array_push($energy,$ingredient->energy*0.1*$portionFactor);
+                array_push($protein,$ingredient->protein*0.1*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.1*$portionFactor);
+                array_push($fat,$ingredient->fat*0.1*$portionFactor);
+            }
+            // Fette 15% proportion
+            if($ingredient->category_id == 5){
+                array_push($energy,$ingredient->energy*0.15*$portionFactor);
+                array_push($protein,$ingredient->protein*0.15*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.15*$portionFactor);
+                array_push($fat,$ingredient->fat*0.15*$portionFactor);
+            }
+            // Gemüse, Kohlenhydrate, Proteine 20% proportion
+            if($ingredient->category_id == 2 || $ingredient->category_id == 3 || $ingredient->category_id == 4){
+                array_push($energy,$ingredient->energy*0.2*$portionFactor);
+                array_push($protein,$ingredient->protein*0.2*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.2*$portionFactor);
+                array_push($fat,$ingredient->fat*0.2*$portionFactor);
+            }
         }
         $energy = array_sum($energy);
         $protein = array_sum($protein);
@@ -66,7 +91,6 @@ class IndexController extends Controller
 
         
         $recipes = Recipe::all();
-        $recipeCount = Recipe::where('user_id',auth()->user()->id)->where('is_bookmarked',false)->count();
        
         foreach($recipes as $recipe){
             // Delete all recipes with user_id = NULL
@@ -74,7 +98,7 @@ class IndexController extends Controller
                 $recipe->delete();
             }
             // Delete last 5 recipes with user_id = auth()->user()->id
-            else{
+            if(isset(auth()->user()->id)){
                 $keep = Recipe::where('user_id',auth()->user()->id)->where('is_bookmarked',false)->latest()->take(4)->pluck('id');
                 Recipe::where('user_id',auth()->user()->id)->where('is_bookmarked',false)->whereNotIn('id', $keep)->delete();
             }
@@ -167,10 +191,51 @@ class IndexController extends Controller
 
     // TODO: doc
     public function pdf(){
-        $recipe = Recipe::latest('created_at')->first()->ingredients()->orderBy('category_id')->get();
+        $ingredients = Recipe::latest('created_at')->first()->ingredients()->orderBy('category_id')->get();
 
-        $pdf = Pdf::loadView('pdf',compact('recipe'));
+        // Energy, Proteins, Carbohydrates, Fat
+        $portionFactor = 2;
+        $energy = [];
+        $protein = [];
+        $carbohydrate = [];
+        $fat = [];
+        foreach($ingredients as $ingredient){
+            // Topping 5% proportion
+            if($ingredient->category_id == 7){
+                array_push($energy,$ingredient->energy*0.05*$portionFactor);
+                array_push($protein,$ingredient->protein*0.05*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.05*$portionFactor);
+                array_push($fat,$ingredient->fat*0.05*$portionFactor);
+            }
+            // Salatbasis & Früchte 10% proportion
+            if($ingredient->category_id == 1 || $ingredient->category_id == 6){
+                array_push($energy,$ingredient->energy*0.1*$portionFactor);
+                array_push($protein,$ingredient->protein*0.1*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.1*$portionFactor);
+                array_push($fat,$ingredient->fat*0.1*$portionFactor);
+            }
+            // Fette 15% proportion
+            if($ingredient->category_id == 5){
+                array_push($energy,$ingredient->energy*0.15*$portionFactor);
+                array_push($protein,$ingredient->protein*0.15*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.15*$portionFactor);
+                array_push($fat,$ingredient->fat*0.15*$portionFactor);
+            }
+            // Gemüse, Kohlenhydrate, Proteine 20% proportion
+            if($ingredient->category_id == 2 || $ingredient->category_id == 3 || $ingredient->category_id == 4){
+                array_push($energy,$ingredient->energy*0.2*$portionFactor);
+                array_push($protein,$ingredient->protein*0.2*$portionFactor);
+                array_push($carbohydrate,$ingredient->carbohydrate*0.2*$portionFactor);
+                array_push($fat,$ingredient->fat*0.2*$portionFactor);
+            }
+        }
+        $energy = array_sum($energy);
+        $protein = array_sum($protein);
+        $carbohydrate = array_sum($carbohydrate);
+        $fat = array_sum($fat);
+
+        $pdf = Pdf::loadView('pdf',compact('ingredients','energy','protein','carbohydrate','fat'));
         return $pdf->download('Rezept.pdf');
-        return view('pdf',compact('recipe')); // TODO: entfernen
+        return view('pdf',compact('ingredients','energy','protein','carbohydrate','fat')); // TODO: entfernen
     }
 }
